tkd/requirements.txt
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.5
dicttoxml==1.7.16
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
fire==0.6.0
flatbuffers==24.3.25
fonttools==4.53.1
futures==3.0.5
goslate==1.5.4
gTTS==2.5.1
h11==0.14.0
humanfriendly==10.0
idna==3.7
imageio==2.34.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lazy_loader==0.4
llvmlite==0.43.0
lxml==5.2.2
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
onnxruntime==1.18.1
opencv-python-headless==4.10.0.84
outcome==1.3.0.post0
packaging==24.1
pdf2docx==0.5.8
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
protobuf==5.27.2
pycparser==2.22
PyDictionary==2.0.1
PyJWT==2.8.0
PyMatting==1.1.12
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pypiwin32==223
pyreadline3==3.4.1
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-whois==0.9.4
pyttsx3==2.90
pytube==15.0.0
pytz==2024.1
pywin32==306
referencing==0.35.1
rembg==2.0.57
requests==2.32.3
rpds-py==0.19.0
scikit-image==0.24.0
scipy==1.14.0
selenium==4.23.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
sympy==1.13.0
termcolor==2.4.0
tifffile==2024.7.2
tqdm==4.66.4
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websocket-client==1.8.0
whois==1.20240129.2
wsproto==1.2.0
xmltodict==0.13.0
bcrypt
moviepy
qrcode
markdown
barcode
html2text
colorthief
PyPDF2
moviepymoviepy==1.0.3